backend/auth/requirements.txt
fastapi
uvicorn
supabase
pydantic[email]
python-dotenv
